Class Code/Class 4/agents-sdk-chainlit/pyproject.toml
[project]
name = "agents-sdk-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.0",
    "openai-agents>=0.1.0",
]
